+++ weatherly/abc.py
@@ -290,8 +290,8 @@ class ForecastHourModel(ResponseModel):
         Wind gust in miles per hour
     gust_kph: :class:`float`
         Wind gust in kilometer per hour
-    uv: :class:`float`
-        UV Index
+    uv: Optional[:class:`float`]
+        UV Index. Can be ``None`` when this class is returned from Future API 
     aqi: Optional[:class:`AirQuality`]
         Air Quality data. See :class:`AirQualityData` for more info.
     """
@@ -329,7 +329,7 @@ class ForecastHourModel(ResponseModel):
     chance_of_snow: int
     gust_mph: float
     gust_kph: float
-    uv: float
+    uv: Optional[float]
 
     aqi: Optional[AirQuality]
     
@@ -373,8 +373,8 @@ class ForecastDayModel(ResponseModel):
         Average visibility in miles
     avghumidity: :class:`int`
         Average humidity as percentage
-    uv: :class:`int`
-        UV Index
+    uv: Optional[:class:`float`]
+        UV Index. Can be ``None`` when returned from Future API
     condition_text: :class:`str`
         Weather condition text
     condition_icon: :class:`str`
@@ -425,7 +425,7 @@ class ForecastDayModel(ResponseModel):
     avgvis_km: float
     avgvis_miles: float
     avghumidity: int
-    uv: int
+    uv: Optional[float]
 
     condition_text: str
     condition_icon: str
@@ -512,221 +512,7 @@ class ForecastModel(ResponseModel):
     location: LocationModel
     forecast_days: List[ForecastDayModel]
     alerts: List[AlertModel]
-
-class FutureHourModel(ResponseModel):
-    """
-    An ABC defining a base future hour model from WeatherAPI
-
-    The following classes implement this ABC:
-    - :class:`~weatherly.FutureHour`
-
-    Attributes
-    -------------
-    time_epoch: :class:`int`
-        Time as epoch
-    time: :class:`str`
-        Date and time
-    temp_c: :class:`float`
-        Temperature in celsius
-    temp_f: :class:`float`
-        Temperature in fahrenheit
-    condition_text: :class:`str`
-        Weather condition text
-    condition_icon: :class:`str`
-        Weather condition icons
-    condition_code: :class:`int`
-        Weather condition code
-    wind_mph: :class:`float`
-        Maximum wind speed in miles per hour
-    wind_kph: :class:`float`
-        Maximum wind speed in kilometer per hour
-    wind_degree: :class:`int`
-        Wind direction in degrees
-    wind_dir: :class:`str`
-        Wind direction as 16 point compass. e.g.: NSW
-    pressure_mb: :class:`float`
-        Pressure in millibars
-    pressure_in: :class:`float`
-        Pressure in inches
-    precip_mm: :class:`float`
-        Precipitation amount in millimeters
-    precip_in: :class:`float`
-        Precipitation amount in inches
-    humidity: :class:`int`
-        Humidity as percentage
-    cloud: :class:`int`
-        Cloud cover as percentage
-    feelslike_c: :class:`float`
-        Feels like temperature as celcius
-    feelslike_f: :class:`float`
-        Feels like temperature as fahrenheit
-    windchill_c: :class:`float`
-        Windchill temperature in celcius
-    windchill_f: :class:`float`
-        Windchill temperature in fahrenheit
-    headindex_c: :class:`float`
-        Heat index in celcius
-    headindex_f: :class:`float`
-        Heat index in fahrenheit
-    dewpoint_c: :class:`float`
-        Dew point in celcius
-    dewpoint_f: :class:`float`
-        Dew point in fahrenheit
-    will_it_rain: :class:`bool`
-        Will it will rain or not
-    will_it_snow: :class:`bool`
-        Will it will snow or not
-    is_day: :class:`bool`
-        Whether to show day condition icon or night icon
-    vis_km: :class:`float`
-        Visibility in kilometer
-    vis_miles: :class:`float`
-        Visibility in miles
-    chance_of_rain: :class:`int`
-        Chance of rain as percentage
-    chance_of_snow: :class:`int`
-        Chance of snow as percentage
-    gust_mph: :class:`float`
-        Wind gust in miles per hour
-    gust_kph: :class:`float`
-        Wind gust in kilometer per hour
-    """
-    time_epoch: int
-    time: str
-    temp_c: float
-    temp_f: float
-    condition_text: str
-    condition_icon: str
-    condition_code: int
-    wind_mph: float
-    wind_kph: float
-    wind_degree: int
-    wind_dir: str
-    pressure_mb: float
-    pressure_in: float
-    precip_mm: float
-    precip_in: float
-    humidity: int
-    cloud: int
-    feelslike_c: float
-    feelslike_f: float
-    windchill_c: float
-    windchill_f: float
-    headindex_c: float
-    headindex_f: float
-    dewpoint_c: float
-    dewpoint_f: float
-    will_it_rain: bool
-    will_it_snow: bool
-    is_day: bool
-    vis_km: float
-    vis_miles: float
-    chance_of_rain: int
-    chance_of_snow: int
-    gust_mph: float
-    gust_kph: float
     
-
-class FutureDayModel(ResponseModel):
-    """
-    An ABC defining a base future day model from WeatherAPI
-
-    The following classes implement this ABC:
-    - :class:`~weatherly.FutureDay`
-
-    Attributes
-    -------------
-    date: :class:`str`
-        Future date
-    date_epoch: :class:`int`
-        Future date as unix time.
-    maxtemp_c: :class:`float`
-        Maximum temperature in celsius for the day.
-    maxtemp_f: :class:`float`
-        Maximum temperature in fahrenheit for the day
-    mintemp_c: :class:`float`
-        Minimum temperature in celsius for the day
-    mintemp_f: :class:`float`
-        Minimum temperature in fahrenheit for the day
-    avgtemp_c: :class:`float`
-        Average temperature in celsius for the day
-    avgtemp_f: :class:`float`
-        Average temperature in fahrenheit for the day
-    maxwind_mph: :class:`float`
-        Maximum wind speed in miles per hour
-    maxwind_mph: :class:`float`
-        Maximum wind speed in kilometer per hour
-    totalprecip_mm: :class:`float`
-        Total precipitation in milimeter
-    totalprecip_in: :class:`float`
-        Total precipitation in inches
-    avgvis_km: :class:`float`
-        Average visibility in kilometer
-    avgvis_miles: :class:`float`
-        Average visibility in miles
-    avghumidity: :class:`int`
-        Average humidity as percentage
-    condition_text: :class:`str`
-        Weather condition text
-    condition_icon: :class:`str`
-        Weather condition icon
-    condition_code: :class:`int`
-        Weather condition code
-    hour_data: List[:class:`FutureHourModel`]
-        A list of :class:`FutureHourModel` objects representing hourly weather data.
-    sunrise: :class:`str`
-        Sunrise time
-    sunset: :class:`str`
-        Sunset time
-    moonrise: :class:`str`
-        Moonrise time
-    moonset: :class:`str`
-        Moonset time
-    moon_phase: :class:`str`
-        Moon phases. Value returned:
-        * New Moon
-        * Waxing Crescent
-        * First Quarter
-        * Waxing Gibbous
-        * Full Moon
-        * Waning Gibbous
-        * Last Quarter
-        * Waning Crescent
-    moon_illumination: :class:`float`
-        Moon illumination as %
-    """
-    date: str
-    date_epoch: int
-    # day
-    maxtemp_c: float
-    maxtemp_f: float
-    mintemp_c: float
-    mintemp_f: float
-    avgtemp_c: float
-    avgtemp_f: float
-
-    maxwind_mph: float
-    maxwind_mph: float
-
-    totalprecip_mm: float
-    totalprecip_in: float
-
-    avgvis_km: float
-    avgvis_miles: float
-    avghumidity: int
-
-    condition_text: str
-    condition_icon: str
-    condition_code: int
-    # astro data
-    sunrise: Optional[str]
-    sunset: Optional[str]
-    moonrise: Optional[str]
-    moonset: Optional[str]
-    moon_phase: Optional[str]
-    moon_illumination: Optional[float]
-
-    hour_data: List[FutureHourModel]
 
 class FutureModel(ResponseModel):
     """
@@ -739,8 +525,8 @@ class FutureModel(ResponseModel):
     -------------
     location: :class:`LocationModel`
         Location of the forecast data.
-    forecast_days: :class:`FutureDayModel`
+    day: :class:`ForecastDayModel`
         Day data
     """
     location: LocationModel
-    day: FutureDayModel
+    day: ForecastDayModel