+++ weatherly/responses.py
@@ -30,8 +30,6 @@ from .abc import (
     ForecastHourModel, 
     ForecastModel,
     AlertModel,
-    FutureDayModel,
-    FutureHourModel,
     FutureModel
 )
 from typing import (
@@ -51,8 +49,6 @@ __all__ = (
     "ForecastDay",
     "ForecastHour",
     "AlertData",
-    "FutureDay",
-    "FutureHour",
     "FutureData",
 )
 
@@ -221,11 +217,11 @@ class AirQualityData(AirQuality):
         Carbon Monoxide (μg/m3)
     o3: :class:`float`
         Ozone (μg/m3)
-    no2: :class`float`	
+    no2: :class:`float`	
         Nitrogen dioxide (μg/m3)
     so2: :class:`float`
         Sulphur dioxide (μg/m3)
-    pm2_5: :class`float`
+    pm2_5: :class:`float`
         PM2.5 (μg/m3)
     pm10: :class:`float`
         PM10 (μg/m3)
@@ -247,8 +243,9 @@ class AirQualityData(AirQuality):
         +--------+------+-------+-------+----------+----------+----------+-------+-------+-------+------------+
         | µgm^-3 | 0-11 | 12-23 | 24-35 | 36-41    | 42-47    | 48-53    | 54-58 | 59-64 | 65-70 | 71 or more |
         +--------+------+-------+-------+----------+----------+----------+-------+-------+-------+------------+
+    
     gb_defra_band: :class:`str`
-        A band corresponding to the :ref:`gb_defra_index`
+        A band corresponding to the :py:attr:`~gb_defra_index`
     """
     def __init__(
         self,
@@ -414,8 +411,8 @@ class ForecastHour(ForecastHourModel):
         Wind gust in miles per hour
     gust_kph: :class:`float`
         Wind gust in kilometer per hour
-    uv: :class:`float`
-        UV Index
+    uv: Optional[:class:`float`]
+        UV Index. Can be ``None`` when this class is returned from Future API
     aqi: Optional[:class:`AirQualityData`]
         Air Quality data. See :class:`AirQualityData` for more info.
     """
@@ -463,7 +460,7 @@ class ForecastHour(ForecastHourModel):
         self.chance_of_snow = raw["chance_of_snow"]
         self.gust_mph = raw["gust_mph"]
         self.gust_kph = raw["gust_kph"]
-        self.uv = raw["uv"]
+        self.uv = raw.get("uv")
         
         if raw.get("air_quality"): self.aqi = AirQualityData(raw["air_quality"], status, code)
         else: self.aqi = None
@@ -510,8 +507,8 @@ class ForecastDay(ForecastDayModel):
         Average visibility in miles
     avghumidity: :class:`int`
         Average humidity as percentage
-    uv: :class:`int`
-        UV Index
+    uv: Optional[:class:`float`]
+        UV Index. Can be ``None`` when this class is returned from Future API
     condition_text: :class:`str`
         Weather condition text
     condition_icon: :class:`str`
@@ -571,7 +568,7 @@ class ForecastDay(ForecastDayModel):
         self.avgvis_km = raw["avgvis_km"]
         self.avgvis_miles = raw["avgvis_miles"]
         self.avghumidity = raw["avghumidity"]
-        self.uv = raw["uv"]
+        self.uv = raw.get("uv")
         self.condition_text = raw["condition"]["text"]
         self.condition_icon = raw["condition"]["icon"]
         self.condition_code = raw["condition"]["code"]
@@ -650,245 +647,6 @@ class ForecastData(ForecastModel):
             for hour in day.hour_data:
                 yield hour
 
-class FutureHour(FutureHourModel):
-    """
-    Future hour, an element of :class:`FutureDay`
-    
-    Attributes
-    -------------
-    raw: Dict[:class:`str`, Any]
-        Raw response in a JSON-like format (converted to a python dictionary)
-    status: :class:`int`
-        HTTP status of the response. 200 is OK, and is the most common status.
-    code: Optional[:class:`int`]
-        Response code. In some cases this can be ``None``
-    time_epoch: :class:`int`
-        Time as epoch
-    time: :class:`str`
-        Date and time
-    temp_c: :class:`float`
-        Temperature in celsius
-    temp_f: :class:`float`
-        Temperature in fahrenheit
-    condition_text: :class:`str`
-        Weather condition text
-    condition_icon: :class:`str`
-        Weather condition icons
-    condition_code: :class:`int`
-        Weather condition code
-    wind_mph: :class:`float`
-        Maximum wind speed in miles per hour
-    wind_kph: :class:`float`
-        Maximum wind speed in kilometer per hour
-    wind_degree: :class:`int`
-        Wind direction in degrees
-    wind_dir: :class:`str`
-        Wind direction as 16 point compass. e.g.: NSW
-    pressure_mb: :class:`float`
-        Pressure in millibars
-    pressure_in: :class:`float`
-        Pressure in inches
-    precip_mm: :class:`float`
-        Precipitation amount in millimeters
-    precip_in: :class:`float`
-        Precipitation amount in inches
-    humidity: :class:`int`
-        Humidity as percentage
-    cloud: :class:`int`
-        Cloud cover as percentage
-    feelslike_c: :class:`float`
-        Feels like temperature as celcius
-    feelslike_f: :class:`float`
-        Feels like temperature as fahrenheit
-    windchill_c: :class:`float`
-        Windchill temperature in celcius
-    windchill_f: :class:`float`
-        Windchill temperature in fahrenheit
-    headindex_c: :class:`float`
-        Heat index in celcius
-    headindex_f: :class:`float`
-        Heat index in fahrenheit
-    dewpoint_c: :class:`float`
-        Dew point in celcius
-    dewpoint_f: :class:`float`
-        Dew point in fahrenheit
-    will_it_rain: :class:`bool`
-        Will it will rain or not
-    will_it_snow: :class:`bool`
-        Will it will snow or not
-    is_day: :class:`bool`
-        Whether to show day condition icon or night icon
-    vis_km: :class:`float`
-        Visibility in kilometer
-    vis_miles: :class:`float`
-        Visibility in miles
-    chance_of_rain: :class:`int`
-        Chance of rain as percentage
-    chance_of_snow: :class:`int`
-        Chance of snow as percentage
-    gust_mph: :class:`float`
-        Wind gust in miles per hour
-    gust_kph: :class:`float`
-        Wind gust in kilometer per hour
-    """
-    def __init__(
-        self,
-        raw: Dict[str, Any],
-        status: int,
-        code: Optional[int]
-    ) -> None:
-        self.raw = raw
-        self.status = status
-        self.code = code
-
-        self.time_epoch = raw["time_epoch"]
-        self.time = raw["time"]
-        self.temp_c = raw["temp_c"]
-        self.temp_f = raw["temp_f"]
-        self.condition_text = raw["condition"]["text"]
-        self.condition_icon = raw["condition"]["icon"]
-        self.condition_code = raw["condition"]["code"]
-        self.wind_mph = raw["wind_mph"]
-        self.wind_kph = raw["wind_kph"]
-        self.wind_degree = raw["wind_degree"]
-        self.wind_dir = raw["wind_dir"]
-        self.pressure_mb = raw["pressure_mb"]
-        self.pressure_in = raw["pressure_in"]
-        self.precip_mm = raw["precip_mm"]
-        self.precip_in = raw["precip_in"]
-        self.humidity = raw["humidity"]
-        self.cloud = raw["cloud"]
-        self.feelslike_c = raw["feelslike_c"]
-        self.feelslike_f = raw["feelslike_f"]
-        self.windchill_c = raw["windchill_c"]
-        self.windchill_f = raw["windchill_f"]
-        self.heatindex_c = raw["heatindex_c"]
-        self.heatindex_f = raw["heatindex_f"]
-        self.dewpoint_c = raw["dewpoint_c"]
-        self.dewpoint_f = raw["dewpoint_f"]
-        self.will_it_rain = bool(raw["will_it_rain"])
-        self.will_it_snow = bool(raw["will_it_snow"])
-        self.is_day = bool(raw["is_day"])
-        self.vis_km = raw["vis_km"]
-        self.vis_miles = raw["vis_miles"]
-        self.chance_of_rain = raw["chance_of_rain"]
-        self.chance_of_snow = raw["chance_of_snow"]
-        self.gust_mph = raw["gust_mph"]
-        self.gust_kph = raw["gust_kph"]
-
-class FutureDay(FutureDayModel):
-    """
-    A future day, element of :class:`FutureData`
-
-    Attributes
-    -------------
-    raw: Dict[:class:`str`, Any]
-        Raw response in a JSON-like format (converted to a python dictionary)
-    status: :class:`int`
-        HTTP status of the response. 200 is OK, and is the most common status.
-    code: Optional[:class:`int`]
-        Response code. In some cases this can be ``None``
-    date: :class:`str`
-        Forecast date
-    date_epoch: :class:`int`
-        Forecast date as unix time.
-    maxtemp_c: :class:`float`
-        Maximum temperature in celsius for the day.
-    maxtemp_f: :class:`float`
-        Maximum temperature in fahrenheit for the day
-    mintemp_c: :class:`float`
-        Minimum temperature in celsius for the day
-    mintemp_f: :class:`float`
-        Minimum temperature in fahrenheit for the day
-    avgtemp_c: :class:`float`
-        Average temperature in celsius for the day
-    avgtemp_f: :class:`float`
-        Average temperature in fahrenheit for the day
-    maxwind_mph: :class:`float`
-        Maximum wind speed in miles per hour
-    maxwind_kph: :class:`float`
-        Maximum wind speed in kilometer per hour
-    totalprecip_mm: :class:`float`
-        Total precipitation in milimeter
-    totalprecip_in: :class:`float`
-        Total precipitation in inches
-    avgvis_km: :class:`float`
-        Average visibility in kilometer
-    avgvis_miles: :class:`float`
-        Average visibility in miles
-    avghumidity: :class:`int`
-        Average humidity as percentage
-    condition_text: :class:`str`
-        Weather condition text
-    condition_icon: :class:`str`
-        Weather condition icon
-    condition_code: :class:`int`
-        Weather condition code
-    hour_data: List[:class:`FutureHour`]
-        A list of :class:`FutureHour` objects representing hourly weather data.
-    sunrise: :class:`str`
-        Sunrise time
-    sunset: :class:`str`
-        Sunset time
-    moonrise: :class:`str`
-        Moonrise time
-    moonset: :class:`str`
-        Moonset time
-    moon_phase: :class:`str`
-        Moon phases. Value returned:
-            * New Moon
-            * Waxing Crescent
-            * First Quarter
-            * Waxing Gibbous
-            * Full Moon
-            * Waning Gibbous
-            * Last Quarter
-            * Waning Crescent
-    moon_illumination: :class:`float`
-        Moon illumination as %
-    """
-    def __init__(
-        self,
-        raw: Dict[str, Any],
-        status: int,
-        code: Optional[int]
-    ) -> None:
-        self.raw = raw
-        self.status = status
-        self.code = code
-        self.date = raw["date"]
-        self.date_epoch = raw["date_epoch"]
-        
-        before_raw = raw.copy()
-        raw = raw["day"]
-        
-        self.maxtemp_c = raw["maxtemp_c"]
-        self.maxtemp_f = raw["maxtemp_f"]
-        self.mintemp_c = raw["mintemp_c"]
-        self.mintemp_f = raw["mintemp_f"]
-        self.avgtemp_c = raw["avgtemp_c"]
-        self.avgtemp_f = raw["avgtemp_f"]
-        self.maxwind_mph = raw["maxwind_mph"]
-        self.maxwind_kph = raw["maxwind_kph"]
-        self.totalprecip_in = raw["totalprecip_in"]
-        self.totalprecip_mm = raw["totalprecip_mm"]
-        self.avgvis_km = raw["avgvis_km"]
-        self.avgvis_miles = raw["avgvis_miles"]
-        self.avghumidity = raw["avghumidity"]
-        self.condition_text = raw["condition"]["text"]
-        self.condition_icon = raw["condition"]["icon"]
-        self.condition_code = raw["condition"]["code"]
-        
-        raw = before_raw
-        self.hour_data = list([
-            FutureHour(elem, status, code) for elem in raw["hour"]
-        ])
-        self.sunrise = raw["astro"]["sunrise"]
-        self.sunset = raw["astro"]["sunset"]
-        self.moonrise = raw["astro"]["moonrise"]
-        self.moonset = raw["astro"]["moonset"]
-        self.moon_phase = raw["astro"]["moon_phase"]
-        self.moon_illumination = raw["astro"]["moon_illumination"]
 
 class FutureData(FutureModel):
     """
@@ -904,7 +662,7 @@ class FutureData(FutureModel):
         Response code. In some cases this can be ``None``
     location: :class:`LocationData`
         Location of the forecast data.
-    day: :class:`FutureDay`
+    day: :class:`ForecastDay`
         Day data for the requested future date.
     """
     def __init__(
@@ -918,4 +676,4 @@ class FutureData(FutureModel):
         self.code = code
         
         self.location = LocationData(raw["location"], status, code)
-        self.day = FutureDay(raw["forecast"]["forecastday"][0], status, code)
+        self.day = ForecastDay(raw["forecast"]["forecastday"][0], status, code)